Add in some files from SVN
There's a later commit in the subversion repo with the label "adding all the
code I wrote" by peter; this is dated 2021-01-04, so later than the latest
commit to this git repo.

Seems to run well enough, but the committed version is of course the last runs,
with maximum detail, so awfully slow.

Copied by hand, so not sure if entirely correct or complete.

GK, 23-May-MMXXIV